HADOOP-373.  Consistently check the value of FileSystem.mkdirs().  Contributed by Wendy.
git-svn-id: https://svn.apache.org/repos/asf/lucene/hadoop/trunk@470195 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/mapred/JobTracker.java
@@ -495,7 +495,9 @@ public class JobTracker implements MRConstants, InterTrackerProtocol, JobSubmiss
         this.systemDir = jobConf.getSystemDir();
         this.fs = FileSystem.get(conf);
         fs.delete(systemDir);
-        fs.mkdirs(systemDir);
+        if (!fs.mkdirs(systemDir)) {
+          throw new IOException("Mkdirs failed to create " + systemDir.toString());
+        }
 
         // Same with 'localDir' except it's always on the local disk.
         jobConf.deleteLocalFiles(SUBDIR);
